refactor: MenuFeature 클래스를 feature 패키지로 이동한다.
